TST: misplaced tests (#56424)
