+++ pandas/tests/io/test_feather.py
@@ -15,7 +15,7 @@ pytestmark = pytest.mark.filterwarnings(
     "ignore:Passing a BlockManager to DataFrame:DeprecationWarning"
 )
 
-pyarrow = pytest.importorskip("pyarrow")
+pa = pytest.importorskip("pyarrow")
 
 
 @pytest.mark.single_cpu
@@ -169,7 +169,6 @@ class TestFeather:
 
     def test_read_feather_dtype_backend(self, string_storage, dtype_backend):
         # GH#50765
-        pa = pytest.importorskip("pyarrow")
         df = pd.DataFrame(
             {
                 "a": pd.Series([1, np.nan, 3], dtype="Int64"),
